Initial implementation of the new tokenizer
