+++ common/issues/data.js
@@ -58,6 +58,11 @@ export default {
     level: 'error',
     message: stringTemplate`Either "${'previousTag'}" contains a comma when it should not or "${'tag'}" is not a valid tag.`,
   },
+  invalidTagPrefix: {
+    hedCode: 'TAG_NAMESPACE_PREFIX_INVALID',
+    level: 'error',
+    message: stringTemplate`Either tag prefix at index ${'index'} contains non-alphabetic characters or does not have an associated schema.`,
+  },
   multipleUniqueTags: {
     hedCode: 'TAG_NOT_UNIQUE',
     level: 'error',
