+++ tests/event.spec.js
@@ -64,22 +64,19 @@ describe('HED string and event validation', () => {
       it('should not have mismatched parentheses', () => {
         const testStrings = {
           extraOpening:
-            '/Action/Reach/To touch,((/Attribute/Object side/Left,/Participant/Effect/Body part/Arm),/Attribute/Location/Screen/Top/70 px,/Attribute/Location/Screen/Left/23 px',
+            'Action/Reach/To touch,((Attribute/Object side/Left,Participant/Effect/Body part/Arm),Attribute/Location/Screen/Top/70 px,Attribute/Location/Screen/Left/23 px',
           // The extra comma is needed to avoid a comma error.
           extraClosing:
-            '/Action/Reach/To touch,(/Attribute/Object side/Left,/Participant/Effect/Body part/Arm),),/Attribute/Location/Screen/Top/70 px,/Attribute/Location/Screen/Left/23 px',
+            'Action/Reach/To touch,(Attribute/Object side/Left,Participant/Effect/Body part/Arm),),Attribute/Location/Screen/Top/70 px,Attribute/Location/Screen/Left/23 px',
           wrongOrder:
-            '/Action/Reach/To touch,((/Attribute/Object side/Left),/Participant/Effect/Body part/Arm),/Attribute/Location/Screen/Top/70 px),(/Attribute/Location/Screen/Left/23 px',
+            'Action/Reach/To touch,((Attribute/Object side/Left),Participant/Effect/Body part/Arm),Attribute/Location/Screen/Top/70 px),(Attribute/Location/Screen/Left/23 px',
           valid:
-            '/Action/Reach/To touch,(/Attribute/Object side/Left,/Participant/Effect/Body part/Arm),/Attribute/Location/Screen/Top/70 px,/Attribute/Location/Screen/Left/23 px',
+            'Action/Reach/To touch,(Attribute/Object side/Left,Participant/Effect/Body part/Arm),Attribute/Location/Screen/Top/70 px,Attribute/Location/Screen/Left/23 px',
         }
         const expectedIssues = {
           extraOpening: [generateIssue('parentheses', { opening: 2, closing: 1 })],
           extraClosing: [generateIssue('parentheses', { opening: 1, closing: 2 })],
-          wrongOrder: [
-            generateIssue('unopenedParenthesis', { index: 125, string: testStrings.wrongOrder }),
-            generateIssue('unclosedParenthesis', { index: 127, string: testStrings.wrongOrder }),
-          ],
+          wrongOrder: [generateIssue('unopenedParenthesis', { index: 121, string: testStrings.wrongOrder })],
           valid: [],
         }
         // No-op function as this check is done during the parsing stage.
@@ -90,31 +87,31 @@ describe('HED string and event validation', () => {
       it('should not have malformed delimiters', () => {
         const testStrings = {
           missingOpeningComma:
-            '/Action/Reach/To touch(/Attribute/Object side/Left,/Participant/Effect/Body part/Arm),/Attribute/Location/Screen/Top/70 px,/Attribute/Location/Screen/Left/23 px',
+            'Action/Reach/To touch(Attribute/Object side/Left,Participant/Effect/Body part/Arm),Attribute/Location/Screen/Top/70 px,Attribute/Location/Screen/Left/23 px',
           missingClosingComma:
-            '/Action/Reach/To touch,(/Attribute/Object side/Left,/Participant/Effect/Body part/Arm)/Attribute/Location/Screen/Top/70 px,/Attribute/Location/Screen/Left/23 px',
+            'Action/Reach/To touch,(Attribute/Object side/Left,Participant/Effect/Body part/Arm)Attribute/Location/Screen/Top/70 px,Attribute/Location/Screen/Left/23 px',
           extraOpeningComma:
-            ',/Action/Reach/To touch,(/Attribute/Object side/Left,/Participant/Effect/Body part/Arm),/Attribute/Location/Screen/Top/70 px,/Attribute/Location/Screen/Left/23 px',
+            ',Action/Reach/To touch,(Attribute/Object side/Left,Participant/Effect/Body part/Arm),Attribute/Location/Screen/Top/70 px,Attribute/Location/Screen/Left/23 px',
           extraClosingComma:
-            '/Action/Reach/To touch,(/Attribute/Object side/Left,/Participant/Effect/Body part/Arm),/Attribute/Location/Screen/Top/70 px,/Attribute/Location/Screen/Left/23 px,',
+            'Action/Reach/To touch,(Attribute/Object side/Left,Participant/Effect/Body part/Arm),/Attribute/Location/Screen/Top/70 px,Attribute/Location/Screen/Left/23 px,',
           multipleExtraOpeningDelimiter:
-            ',,/Action/Reach/To touch,(/Attribute/Object side/Left,/Participant/Effect/Body part/Arm),/Attribute/Location/Screen/Top/70 px,/Attribute/Location/Screen/Left/23 px',
+            ',,Action/Reach/To touch,(Attribute/Object side/Left,Participant/Effect/Body part/Arm),Attribute/Location/Screen/Top/70 px,Attribute/Location/Screen/Left/23 px',
           multipleExtraClosingDelimiter:
-            '/Action/Reach/To touch,(/Attribute/Object side/Left,/Participant/Effect/Body part/Arm),/Attribute/Location/Screen/Top/70 px,/Attribute/Location/Screen/Left/23 px,,',
+            'Action/Reach/To touch,(Attribute/Object side/Left,Participant/Effect/Body part/Arm),/Attribute/Location/Screen/Top/70 px,Attribute/Location/Screen/Left/23 px,,',
           multipleExtraMiddleDelimiter:
-            '/Action/Reach/To touch,,(/Attribute/Object side/Left,/Participant/Effect/Body part/Arm),/Attribute/Location/Screen/Top/70 px,,/Attribute/Location/Screen/Left/23 px',
+            'Action/Reach/To touch,,(Attribute/Object side/Left,Participant/Effect/Body part/Arm),Attribute/Location/Screen/Top/70 px,,Attribute/Location/Screen/Left/23 px',
           valid:
-            '/Action/Reach/To touch,(/Attribute/Object side/Left,/Participant/Effect/Body part/Arm),/Attribute/Location/Screen/Top/70 px,/Attribute/Location/Screen/Left/23 px',
+            'Action/Reach/To touch,(Attribute/Object side/Left,Participant/Effect/Body part/Arm),Attribute/Location/Screen/Top/70 px,Attribute/Location/Screen/Left/23 px',
           validDoubleOpeningParentheses:
-            '/Action/Reach/To touch,((/Attribute/Object side/Left,/Participant/Effect/Body part/Arm),/Attribute/Location/Screen/Top/70 px,/Attribute/Location/Screen/Left/23 px),Event/Duration/3 ms',
+            'Action/Reach/To touch,((Attribute/Object side/Left,Participant/Effect/Body part/Arm),Attribute/Location/Screen/Top/70 px,Attribute/Location/Screen/Left/23 px),Event/Duration/3 ms',
           validDoubleClosingParentheses:
-            '/Action/Reach/To touch,(/Attribute/Object side/Left,/Participant/Effect/Body part/Arm,(/Attribute/Location/Screen/Top/70 px,/Attribute/Location/Screen/Left/23 px)),Event/Duration/3 ms',
+            'Action/Reach/To touch,(Attribute/Object side/Left,Participant/Effect/Body part/Arm,(Attribute/Location/Screen/Top/70 px,Attribute/Location/Screen/Left/23 px)),Event/Duration/3 ms',
         }
         const expectedIssues = {
-          missingOpeningComma: [generateIssue('commaMissing', { tag: '/Action/Reach/To touch(' })],
+          missingOpeningComma: [generateIssue('commaMissing', { tag: 'Action/Reach/To touch(' })],
           missingClosingComma: [
             generateIssue('commaMissing', {
-              tag: '/Participant/Effect/Body part/Arm)',
+              tag: 'Participant/Effect/Body part/Arm)',
             }),
           ],
           extraOpeningComma: [
@@ -158,12 +155,12 @@ describe('HED string and event validation', () => {
           multipleExtraMiddleDelimiter: [
             generateIssue('extraDelimiter', {
               character: ',',
-              index: 23,
+              index: 22,
               string: testStrings.multipleExtraMiddleDelimiter,
             }),
             generateIssue('extraDelimiter', {
               character: ',',
-              index: 125,
+              index: 121,
               string: testStrings.multipleExtraMiddleDelimiter,
             }),
           ],
@@ -178,68 +175,68 @@ describe('HED string and event validation', () => {
 
       it('should not have invalid characters', () => {
         const testStrings = {
-          openingBrace: '/Attribute/Object side/Left,/Participant/Effect{/Body part/Arm',
-          closingBrace: '/Attribute/Object side/Left,/Participant/Effect}/Body part/Arm',
-          openingBracket: '/Attribute/Object side/Left,/Participant/Effect[/Body part/Arm',
-          closingBracket: '/Attribute/Object side/Left,/Participant/Effect]/Body part/Arm',
-          tilde: '/Attribute/Object side/Left,/Participant/Effect~/Body part/Arm',
-          doubleQuote: '/Attribute/Object side/Left,/Participant/Effect"/Body part/Arm',
-          null: '/Attribute/Object side/Left,/Participant/Effect/Body part/Arm\0',
-          tab: '/Attribute/Object side/Left,/Participant/Effect/Body part/Arm\t',
+          openingBrace: 'Attribute/Object side/Left,Participant/Effect{Body part/Arm',
+          closingBrace: 'Attribute/Object side/Left,Participant/Effect}/Body part/Arm',
+          openingBracket: 'Attribute/Object side/Left,Participant/Effect[Body part/Arm',
+          closingBracket: 'Attribute/Object side/Left,Participant/Effect]Body part/Arm',
+          tilde: 'Attribute/Object side/Left,Participant/Effect~/Body part/Arm',
+          doubleQuote: 'Attribute/Object side/Left,Participant/Effect"/Body part/Arm',
+          null: 'Attribute/Object side/Left,Participant/Effect/Body part/Arm\0',
+          tab: 'Attribute/Object side/Left,Participant/Effect/Body part/Arm\t',
         }
         const expectedIssues = {
           openingBrace: [
             generateIssue('invalidCharacter', {
               character: 'LEFT CURLY BRACKET',
-              index: 47,
+              index: 45,
               string: testStrings.openingBrace,
             }),
           ],
           closingBrace: [
             generateIssue('unopenedCurlyBrace', {
-              index: 47,
+              index: 45,
               string: testStrings.closingBrace,
             }),
           ],
           openingBracket: [
             generateIssue('invalidCharacter', {
               character: 'LEFT SQUARE BRACKET',
-              index: 47,
+              index: 45,
               string: testStrings.openingBracket,
             }),
           ],
           closingBracket: [
             generateIssue('invalidCharacter', {
               character: 'RIGHT SQUARE BRACKET',
-              index: 47,
+              index: 45,
               string: testStrings.closingBracket,
             }),
           ],
           tilde: [
             generateIssue('invalidCharacter', {
               character: 'TILDE',
-              index: 47,
+              index: 45,
               string: testStrings.tilde,
             }),
           ],
           doubleQuote: [
             generateIssue('invalidCharacter', {
               character: 'QUOTATION MARK',
-              index: 47,
+              index: 45,
               string: testStrings.doubleQuote,
             }),
           ],
           null: [
             generateIssue('invalidCharacter', {
               character: 'NULL',
-              index: 61,
+              index: 59,
               string: testStrings.null,
             }),
           ],
           tab: [
             generateIssue('invalidCharacter', {
               character: 'CHARACTER TABULATION',
-              index: 61,
+              index: 59,
               string: testStrings.tab,
             }),
           ],
@@ -346,491 +343,6 @@ describe('HED string and event validation', () => {
         }
         validatorSyntactic(testStrings, expectedIssues, (validator, tagLevel) => {
           validator.checkForDuplicateTags(tagLevel)
-        })
-      })
-    })
-  })
-
-  describe('HED-2G validation', () => {
-    describe('Later HED-2G schemas', () => {
-      const hedSchemaFile = 'tests/data/HED7.1.1.xml'
-      let hedSchemas
-
-      beforeAll(async () => {
-        const spec1 = new SchemaSpec('', '7.1.1', '', hedSchemaFile)
-        const specs = new SchemasSpec().addSchemaSpec(spec1)
-        hedSchemas = await buildSchemas(specs)
-      })
-
-      /**
-       * HED 2 semantic validation base function.
-       *
-       * This base function uses the HED 2-specific {@link Hed2Validator} validator class.
-       *
-       * @param {Object<string, string>} testStrings A mapping of test strings.
-       * @param {Object<string, Issue[]>} expectedIssues The expected issues for each test string.
-       * @param {function(HedValidator): void} testFunction A test-specific function that executes the required validation check.
-       * @param {Object<string, boolean>?} testOptions Any needed custom options for the validator.
-       */
-      const validatorSemanticBase = function (testStrings, expectedIssues, testFunction, testOptions = {}) {
-        validatorBase(hedSchemas, Hed2Validator, testStrings, expectedIssues, testFunction, testOptions)
-      }
-
-      describe('Full HED Strings', () => {
-        const validatorSemantic = validatorSemanticBase
-
-        // TODO: Rewrite as HED 3 test
-        it.skip('should not validate strings with extensions that are valid node names', () => {
-          const testStrings = {
-            // Event/Duration/20 cm is an obviously invalid tag that should not be caught due to the first error.
-            red: 'Attribute/Red, Event/Duration/20 cm',
-            redAndBlue: 'Attribute/Red, Attribute/Blue, Event/Duration/20 cm',
-          }
-          const expectedIssues = {
-            red: [
-              generateIssue('invalidParentNode', {
-                tag: 'Red',
-                parentTag: 'Attribute/Visual/Color/Red',
-              }),
-            ],
-            redAndBlue: [
-              generateIssue('invalidParentNode', {
-                tag: 'Red',
-                parentTag: 'Attribute/Visual/Color/Red',
-              }),
-              generateIssue('invalidParentNode', {
-                tag: 'Blue',
-                parentTag: 'Attribute/Visual/Color/Blue',
-              }),
-            ],
-          }
-          // This is a no-op function since this is checked during string parsing.
-          return validatorSemantic(
-            testStrings,
-            expectedIssues,
-            // eslint-disable-next-line no-unused-vars
-            (validator) => {},
-          )
-        })
-      })
-
-      describe('Individual HED Tags', () => {
-        /**
-         * HED 2 individual tag semantic validation base function.
-         *
-         * @param {Object<string, string>} testStrings A mapping of test strings.
-         * @param {Object<string, Issue[]>} expectedIssues The expected issues for each test string.
-         * @param {function(HedValidator, ParsedHedTag, ParsedHedTag): void} testFunction A test-specific function that executes the required validation check.
-         * @param {Object<string, boolean>?} testOptions Any needed custom options for the validator.
-         */
-        const validatorSemantic = function (testStrings, expectedIssues, testFunction, testOptions) {
-          return validatorSemanticBase(
-            testStrings,
-            expectedIssues,
-            (validator) => {
-              let previousTag = new ParsedHedTag('', '', [0, 0], validator.hedSchemas)
-              for (const tag of validator.parsedString.tags) {
-                testFunction(validator, tag, previousTag)
-                previousTag = tag
-              }
-            },
-            testOptions,
-          )
-        }
-
-        it('should exist in the schema or be an allowed extension', () => {
-          const testStrings = {
-            takesValue: 'Event/Duration/3 ms',
-            full: 'Attribute/Object side/Left',
-            extensionAllowed: 'Item/Object/Person/Driver',
-            leafExtension: 'Event/Category/Initial context/Something',
-            nonExtensionAllowed: 'Event/Nonsense',
-            illegalComma: 'Event/Label/This is a label,This/Is/A/Tag',
-            placeholder: 'Item/Object/#',
-          }
-          const expectedIssues = {
-            takesValue: [],
-            full: [],
-            extensionAllowed: [generateIssue('extension', { tag: testStrings.extensionAllowed })],
-            leafExtension: [generateIssue('invalidTag', { tag: testStrings.leafExtension })],
-            nonExtensionAllowed: [
-              generateIssue('invalidTag', {
-                tag: testStrings.nonExtensionAllowed,
-              }),
-            ],
-            illegalComma: [
-              generateIssue('extraCommaOrInvalid', {
-                previousTag: 'Event/Label/This is a label',
-                tag: 'This/Is/A/Tag',
-              }),
-            ],
-            placeholder: [
-              generateIssue('invalidTag', {
-                tag: testStrings.placeholder,
-              }),
-            ],
-          }
-          return validatorSemantic(
-            testStrings,
-            expectedIssues,
-            (validator, tag, previousTag) => {
-              validator.checkIfTagIsValid(tag, previousTag)
-            },
-            { checkForWarnings: true },
-          )
-        })
-
-        it('should have a child when required', () => {
-          const testStrings = {
-            hasChild: 'Event/Category/Experimental stimulus',
-            missingChild: 'Event/Category',
-          }
-          const expectedIssues = {
-            hasChild: [],
-            missingChild: [generateIssue('childRequired', { tag: testStrings.missingChild })],
-          }
-          return validatorSemantic(
-            testStrings,
-            expectedIssues,
-            // eslint-disable-next-line no-unused-vars
-            (validator, tag, previousTag) => {
-              validator.checkIfTagRequiresChild(tag)
-            },
-            { checkForWarnings: true },
-          )
-        })
-
-        it('should have a proper unit when required', () => {
-          const testStrings = {
-            correctUnit: 'Event/Duration/3 ms',
-            correctUnitScientific: 'Event/Duration/3.5e1 ms',
-            correctSingularUnit: 'Event/Duration/1 millisecond',
-            correctPluralUnit: 'Event/Duration/3 milliseconds',
-            correctNoPluralUnit: 'Attribute/Temporal rate/3 hertz',
-            correctPrefixUnit: 'Participant/Effect/Cognitive/Reward/$19.69',
-            correctNonSymbolCapitalizedUnit: 'Event/Duration/3 MilliSeconds',
-            correctSymbolCapitalizedUnit: 'Attribute/Temporal rate/3 kHz',
-            missingRequiredUnit: 'Event/Duration/3',
-            incorrectUnit: 'Event/Duration/3 cm',
-            incorrectNonNumericValue: 'Event/Duration/A ms',
-            incorrectPluralUnit: 'Attribute/Temporal rate/3 hertzs',
-            incorrectSymbolCapitalizedUnit: 'Attribute/Temporal rate/3 hz',
-            incorrectSymbolCapitalizedUnitModifier: 'Attribute/Temporal rate/3 KHz',
-            incorrectNonSIUnitModifier: 'Event/Duration/1 millihour',
-            incorrectNonSIUnitSymbolModifier: 'Attribute/Path/Velocity/100 Mkph',
-            notRequiredNumber: 'Attribute/Visual/Color/Red/0.5',
-            notRequiredScientific: 'Attribute/Visual/Color/Red/5e-1',
-            properTime: 'Item/2D shape/Clock face/08:30',
-            invalidTime: 'Item/2D shape/Clock face/54:54',
-          }
-          const legalTimeUnits = ['s', 'second', 'day', 'minute', 'hour']
-          const legalFrequencyUnits = ['Hz', 'hertz']
-          const legalSpeedUnits = ['m-per-s', 'kph', 'mph']
-          const expectedIssues = {
-            correctUnit: [],
-            correctUnitScientific: [],
-            correctSingularUnit: [],
-            correctPluralUnit: [],
-            correctNoPluralUnit: [],
-            correctPrefixUnit: [],
-            correctNonSymbolCapitalizedUnit: [],
-            correctSymbolCapitalizedUnit: [],
-            missingRequiredUnit: [
-              generateIssue('unitClassDefaultUsed', {
-                defaultUnit: 's',
-                tag: testStrings.missingRequiredUnit,
-              }),
-            ],
-            incorrectUnit: [
-              generateIssue('unitClassInvalidUnit', {
-                tag: testStrings.incorrectUnit,
-                unitClassUnits: legalTimeUnits.sort().join(','),
-              }),
-            ],
-            incorrectNonNumericValue: [
-              generateIssue('invalidValue', {
-                tag: testStrings.incorrectNonNumericValue,
-              }),
-            ],
-            incorrectPluralUnit: [
-              generateIssue('unitClassInvalidUnit', {
-                tag: testStrings.incorrectPluralUnit,
-                unitClassUnits: legalFrequencyUnits.sort().join(','),
-              }),
-            ],
-            incorrectSymbolCapitalizedUnit: [
-              generateIssue('unitClassInvalidUnit', {
-                tag: testStrings.incorrectSymbolCapitalizedUnit,
-                unitClassUnits: legalFrequencyUnits.sort().join(','),
-              }),
-            ],
-            incorrectSymbolCapitalizedUnitModifier: [
-              generateIssue('unitClassInvalidUnit', {
-                tag: testStrings.incorrectSymbolCapitalizedUnitModifier,
-                unitClassUnits: legalFrequencyUnits.sort().join(','),
-              }),
-            ],
-            incorrectNonSIUnitModifier: [
-              generateIssue('unitClassInvalidUnit', {
-                tag: testStrings.incorrectNonSIUnitModifier,
-                unitClassUnits: legalTimeUnits.sort().join(','),
-              }),
-            ],
-            incorrectNonSIUnitSymbolModifier: [
-              generateIssue('unitClassInvalidUnit', {
-                tag: testStrings.incorrectNonSIUnitSymbolModifier,
-                unitClassUnits: legalSpeedUnits.sort().join(','),
-              }),
-            ],
-            notRequiredNumber: [],
-            notRequiredScientific: [],
-            properTime: [],
-            invalidTime: [
-              generateIssue('invalidValue', {
-                tag: testStrings.invalidTime,
-              }),
-            ],
-          }
-          return validatorSemantic(
-            testStrings,
-            expectedIssues,
-            // eslint-disable-next-line no-unused-vars
-            (validator, tag, previousTag) => {
-              validator.checkIfTagUnitClassUnitsAreValid(tag)
-            },
-            { checkForWarnings: true },
-          )
-        })
-      })
-
-      describe('HED Tag Levels', () => {
-        /**
-         * HED 2 Tag level semantic validation base function.
-         *
-         * @param {Object<string, string>} testStrings A mapping of test strings.
-         * @param {Object<string, Issue[]>} expectedIssues The expected issues for each test string.
-         * @param {function(HedValidator, ParsedHedSubstring[]): void} testFunction A test-specific function that executes the required validation check.
-         * @param {Object<string, boolean>?} testOptions Any needed custom options for the validator.
-         */
-        const validatorSemantic = function (testStrings, expectedIssues, testFunction, testOptions = {}) {
-          return validatorSemanticBase(
-            testStrings,
-            expectedIssues,
-            (validator) => {
-              for (const tagGroup of validator.parsedString.tagGroups) {
-                for (const subGroup of tagGroup.subGroupArrayIterator()) {
-                  testFunction(validator, subGroup)
-                }
-              }
-              testFunction(validator, validator.parsedString.parseTree)
-            },
-            testOptions,
-          )
-        }
-
-        it('should not have multiple copies of a unique tag', () => {
-          const testStrings = {
-            legal:
-              'Event/Description/Rail vehicles,Item/Object/Vehicle/Train,(Item/Object/Vehicle/Train,Event/Category/Experimental stimulus)',
-            multipleDesc:
-              'Event/Description/Rail vehicles,Event/Description/Locomotive-pulled or multiple units,Item/Object/Vehicle/Train,(Item/Object/Vehicle/Train,Event/Category/Experimental stimulus)',
-          }
-          const expectedIssues = {
-            legal: [],
-            multipleDesc: [generateIssue('multipleUniqueTags', { tag: 'event/description' })],
-          }
-          return validatorSemantic(testStrings, expectedIssues, (validator, tagLevel) => {
-            validator.checkForMultipleUniqueTags(tagLevel)
-          })
-        })
-      })
-
-      describe('Top-level Tags', () => {
-        const validatorSemantic = validatorSemanticBase
-
-        it('should include all required tags', () => {
-          const testStrings = {
-            complete:
-              'Event/Label/Bus,Event/Category/Experimental stimulus,Event/Description/Shown a picture of a bus,Item/Object/Vehicle/Bus',
-            missingLabel:
-              'Event/Category/Experimental stimulus,Event/Description/Shown a picture of a bus,Item/Object/Vehicle/Bus',
-            missingCategory: 'Event/Label/Bus,Event/Description/Shown a picture of a bus,Item/Object/Vehicle/Bus',
-            missingDescription: 'Event/Label/Bus,Event/Category/Experimental stimulus,Item/Object/Vehicle/Bus',
-            missingAllRequired: 'Item/Object/Vehicle/Bus',
-          }
-          const expectedIssues = {
-            complete: [],
-            missingLabel: [
-              generateIssue('requiredPrefixMissing', {
-                tagPrefix: 'event/label',
-              }),
-            ],
-            missingCategory: [
-              generateIssue('requiredPrefixMissing', {
-                tagPrefix: 'event/category',
-              }),
-            ],
-            missingDescription: [
-              generateIssue('requiredPrefixMissing', {
-                tagPrefix: 'event/description',
-              }),
-            ],
-            missingAllRequired: [
-              generateIssue('requiredPrefixMissing', {
-                tagPrefix: 'event/label',
-              }),
-              generateIssue('requiredPrefixMissing', {
-                tagPrefix: 'event/category',
-              }),
-              generateIssue('requiredPrefixMissing', {
-                tagPrefix: 'event/description',
-              }),
-            ],
-          }
-          return validatorSemantic(
-            testStrings,
-            expectedIssues,
-            (validator) => {
-              validator.checkForRequiredTags()
-            },
-            { checkForWarnings: true },
-          )
-        })
-      })
-    })
-
-    describe('Pre-v7.1.0 HED schemas', () => {
-      const hedSchemaFile = 'tests/data/HED7.0.4.xml'
-      let hedSchemas
-
-      beforeAll(async () => {
-        const spec2 = new SchemaSpec('', '7.0.4', '', hedSchemaFile)
-        const specs = new SchemasSpec().addSchemaSpec(spec2)
-        hedSchemas = await buildSchemas(specs)
-      })
-
-      /**
-       * HED 2 semantic validation base function.
-       *
-       * This base function uses the HED 2-specific {@link Hed2Validator} validator class.
-       *
-       * @param {Object<string, string>} testStrings A mapping of test strings.
-       * @param {Object<string, Issue[]>} expectedIssues The expected issues for each test string.
-       * @param {function(HedValidator): void} testFunction A test-specific function that executes the required validation check.
-       * @param {Object<string, boolean>?} testOptions Any needed custom options for the validator.
-       */
-      const validatorSemanticBase = function (testStrings, expectedIssues, testFunction, testOptions = {}) {
-        validatorBase(hedSchemas, Hed2Validator, testStrings, expectedIssues, testFunction, testOptions)
-      }
-
-      describe('Individual HED Tags', () => {
-        /**
-         * HED 2 individual tag semantic validation base function.
-         *
-         * @param {Object<string, string>} testStrings A mapping of test strings.
-         * @param {Object<string, Issue[]>} expectedIssues The expected issues for each test string.
-         * @param {function(HedValidator, ParsedHedTag, ParsedHedTag): void} testFunction A test-specific function that executes the required validation check.
-         * @param {Object<string, boolean>?} testOptions Any needed custom options for the validator.
-         */
-        const validatorSemantic = function (testStrings, expectedIssues, testFunction, testOptions) {
-          return validatorSemanticBase(
-            testStrings,
-            expectedIssues,
-            (validator) => {
-              let previousTag = new ParsedHedTag('', '', [0, 0], validator.hedSchemas)
-              for (const tag of validator.parsedString.tags) {
-                testFunction(validator, tag, previousTag)
-                previousTag = tag
-              }
-            },
-            testOptions,
-          )
-        }
-
-        it('should have a proper unit when required', () => {
-          const testStrings = {
-            correctUnit: 'Event/Duration/3 ms',
-            correctUnitWord: 'Event/Duration/3 milliseconds',
-            correctUnitScientific: 'Event/Duration/3.5e1 ms',
-            missingRequiredUnit: 'Event/Duration/3',
-            incorrectUnit: 'Event/Duration/3 cm',
-            incorrectNonNumericValue: 'Event/Duration/A ms',
-            incorrectUnitWord: 'Event/Duration/3 nanoseconds',
-            incorrectModifier: 'Event/Duration/3 ns',
-            notRequiredNumber: 'Attribute/Visual/Color/Red/0.5',
-            notRequiredScientific: 'Attribute/Visual/Color/Red/5e-1',
-            properTime: 'Item/2D shape/Clock face/08:30',
-            invalidTime: 'Item/2D shape/Clock face/54:54',
-          }
-          const legalTimeUnits = [
-            's',
-            'second',
-            'seconds',
-            'centiseconds',
-            'centisecond',
-            'cs',
-            'hour:min',
-            'day',
-            'days',
-            'ms',
-            'milliseconds',
-            'millisecond',
-            'minute',
-            'minutes',
-            'hour',
-            'hours',
-          ]
-          const expectedIssues = {
-            correctUnit: [],
-            correctUnitWord: [],
-            correctUnitScientific: [],
-            missingRequiredUnit: [
-              generateIssue('unitClassDefaultUsed', {
-                defaultUnit: 's',
-                tag: testStrings.missingRequiredUnit,
-              }),
-            ],
-            incorrectUnit: [
-              generateIssue('unitClassInvalidUnit', {
-                tag: testStrings.incorrectUnit,
-                unitClassUnits: legalTimeUnits.sort().join(','),
-              }),
-            ],
-            incorrectNonNumericValue: [
-              generateIssue('invalidValue', {
-                tag: testStrings.incorrectNonNumericValue,
-              }),
-            ],
-            incorrectUnitWord: [
-              generateIssue('unitClassInvalidUnit', {
-                tag: testStrings.incorrectUnitWord,
-                unitClassUnits: legalTimeUnits.sort().join(','),
-              }),
-            ],
-            incorrectModifier: [
-              generateIssue('unitClassInvalidUnit', {
-                tag: testStrings.incorrectModifier,
-                unitClassUnits: legalTimeUnits.sort().join(','),
-              }),
-            ],
-            notRequiredNumber: [],
-            notRequiredScientific: [],
-            properTime: [],
-            invalidTime: [
-              generateIssue('invalidValue', {
-                tag: testStrings.invalidTime,
-              }),
-            ],
-          }
-          return validatorSemantic(
-            testStrings,
-            expectedIssues,
-            // eslint-disable-next-line no-unused-vars
-            (validator, tag, previousTag) => {
-              validator.checkIfTagUnitClassUnitsAreValid(tag)
-            },
-            { checkForWarnings: true },
-          )
         })
       })
     })
