+++ tests/temp.spec.js
@@ -19,7 +19,7 @@ import { HedStringTokenizerFirstTry } from '../parser/tokenizerFirstTry'
 describe('HED string parsing', () => {
   it('should include each group as its own single element', () => {
     //const hedString =
-    'Action/Move/Flex,(Relation/Spatial-relation/Left-side-of,Action/Move/Bend,Upper-extremity/Elbow),Position/X-position/70 px,Position/Y-position/23 px'
+    //'Action/Move/Flex,(Relation/Spatial-relation/Left-side-of,Action/Move/Bend,Upper-extremity/Elbow),Position/X-position/70 px,Position/Y-position/23 px'
     //const hedString = 'x/y w/z'
     //const hedString = '(r,z)'
     //const hedString = 'r,'
@@ -29,7 +29,7 @@ describe('HED string parsing', () => {
     //const hedString = '/x'
     //const hedString = 'x/ /y'
     //const hedString = 'x/'
-    const hedString = 'x'
+    const hedString = 'x:z'
     const tok = new HedStringTokenizer(hedString)
     const [tagSpecs, groupBounds, tokenizingIssues] = tok.tokenize()
     assert.isEmpty(Object.values(tokenizingIssues).flat(), 'Parsing issues occurred')
