+++ tests/tokenizerPassingData.js
@@ -35,10 +35,45 @@ export const passingTests = [
         tagSpecs: [new TagSpec('x/y/z', 0, 5, '')],
         groupSpec: new GroupSpec(0, 5, []),
       },
+      {
+        name: 'tag-in-column-spec',
+        string: '{xy}',
+        explanation: 'Single column spec',
+        tagSpecs: [new ColumnSpliceSpec('xy', 0, 3, '')],
+        groupSpec: new GroupSpec(0, 4, []),
+      },
+      {
+        name: 'tag-in-column-spec-multiple-blanks',
+        string: '  { xy  } ',
+        explanation: 'Single column spec with multiple blanks',
+        tagSpecs: [new ColumnSpliceSpec('xy', 2, 8, '')],
+        groupSpec: new GroupSpec(0, 10, []),
+      },
+      {
+        name: 'tag-with-colons-no-blanks',
+        string: 'xy:wz',
+        explanation: 'Tag with a single colon and no blanks',
+        tagSpecs: [new TagSpec('wz', 3, 5, 'xy')],
+        groupSpec: new GroupSpec(0, 5, []),
+      },
+      {
+        name: 'tag-with-multiple-colons',
+        string: 'xy:wz x:y',
+        explanation: 'Tag with one colon marking library and another as part of a value',
+        tagSpecs: [new TagSpec('wz x:y', 3, 9, 'xy')],
+        groupSpec: new GroupSpec(0, 9, []),
+      },
+      {
+        name: 'tags-with-one-value column',
+        string: 'xy x:y',
+        explanation: 'Tag with one colon as part of a value',
+        tagSpecs: [new TagSpec('xy x:y', 0, 6, '')],
+        groupSpec: new GroupSpec(0, 6, []),
+      },
     ],
   },
   {
-    name: 'valid-tags-no-groups',
+    name: 'multiple-tags-no-groups',
     description: 'multiple tags with no groups.',
     warning: false,
     tests: [
@@ -48,6 +83,13 @@ export const passingTests = [
         explanation: 'Multiple tags with no blanks',
         tagSpecs: [new TagSpec('xy', 0, 2, ''), new TagSpec('zy', 3, 5, ''), new TagSpec('wy', 6, 8, '')],
         groupSpec: new GroupSpec(0, 8, []),
+      },
+      {
+        name: 'multiple-tags-with-blanks',
+        string: ' xy,  zy , wy  ',
+        explanation: 'Can have extra blanks',
+        tagSpecs: [new TagSpec('xy', 1, 3, ''), new TagSpec('zy', 6, 8, ''), new TagSpec('wy', 11, 13, '')],
+        groupSpec: new GroupSpec(0, 15, []),
       },
       {
         name: 'multiple-tags-with-blanks',
@@ -105,4 +147,98 @@ export const passingTests = [
       },
     ],
   },
+  {
+    name: 'Complex nested groups',
+    description: 'Groups with complex nesting',
+    warning: false,
+    tests: [
+      {
+        name: 'Single-multi-nested-group',
+        string: '(((xy)))',
+        explanation: 'Single group with deep nesting',
+        tagSpecs: [[[[new TagSpec('xy', 3, 5, '')]]]],
+        groupSpec: new GroupSpec(0, 8, [new GroupSpec(0, 8, [new GroupSpec(1, 7, [new GroupSpec(2, 6, [])])])]),
+      },
+      {
+        name: 'Nested-group-with-splice',
+        string: '((({xy})))',
+        explanation: 'A tag after a group.',
+        tagSpecs: [[[[new ColumnSpliceSpec('xy', 3, 6)]]]],
+        groupSpec: new GroupSpec(0, 10, [new GroupSpec(0, 10, [new GroupSpec(1, 9, [new GroupSpec(2, 8, [])])])]),
+      },
+      {
+        name: 'multiple-tags-in-group',
+        string: '(x,y)',
+        explanation: 'Multiple tags in one group.',
+        tagSpecs: [[new TagSpec('x', 1, 2, ''), new TagSpec('y', 3, 4, '')]],
+        groupSpec: new GroupSpec(0, 5, [new GroupSpec(0, 5, [])]),
+      },
+      {
+        name: 'multiple-unnested-groups',
+        string: 'q, (xy), (zw, uv), p',
+        explanation: 'Multiple unnested tag groups and tags.',
+        tagSpecs: [
+          new TagSpec('q', 0, 1, ''),
+          [new TagSpec('xy', 4, 6, '')],
+          [new TagSpec('zw', 10, 12, ''), new TagSpec('uv', 14, 16, '')],
+          new TagSpec('p', 19, 20, ''),
+        ],
+        groupSpec: new GroupSpec(0, 20, [new GroupSpec(3, 7, []), new GroupSpec(9, 17, [])]),
+      },
+      {
+        name: 'tag-after-group',
+        string: 'x/y,(r,v)',
+        explanation: 'A tag after a group.',
+        tagSpecs: [new TagSpec('x/y', 0, 3, ''), [new TagSpec('r', 5, 6, ''), new TagSpec('v', 7, 8, '')]],
+        groupSpec: new GroupSpec(0, 9, [new GroupSpec(4, 9, [])]),
+      },
+    ],
+  },
+  // {
+  //   name: 'Complex nested groups',
+  //   description: 'Groups with complex nesting',
+  //   warning: false,
+  //   tests: [
+  //     {
+  //       name: 'Single-nested-group',
+  //       string: '(xy)',
+  //       explanation: 'Single group',
+  //       tagSpecs: [[new TagSpec('xy', 1, 3, '')]],
+  //       groupSpec: new GroupSpec(0, 4, [new GroupSpec(0, 4, [])]),
+  //     },
+  //     {
+  //       name: 'tag-after-group',
+  //       string: '(x), p',
+  //       explanation: 'A tag after a group.',
+  //       tagSpecs: [[new TagSpec('x', 1, 2, '')], new TagSpec('p', 5, 6, '')],
+  //       groupSpec: new GroupSpec(0, 6, [new GroupSpec(0, 3, [])]),
+  //     },
+  //     {
+  //       name: 'multiple-tags-in-group',
+  //       string: '(x,y)',
+  //       explanation: 'Multiple tags in one group.',
+  //       tagSpecs: [[new TagSpec('x', 1, 2, ''), new TagSpec('y', 3, 4, '')]],
+  //       groupSpec: new GroupSpec(0, 5, [new GroupSpec(0, 5, [])]),
+  //     },
+  //     {
+  //       name: 'multiple-unnested-groups',
+  //       string: 'q, (xy), (zw, uv), p',
+  //       explanation: 'Multiple unnested tag groups and tags.',
+  //       tagSpecs: [
+  //         new TagSpec('q', 0, 1, ''),
+  //         [new TagSpec('xy', 4, 6, '')],
+  //         [new TagSpec('zw', 10, 12, ''), new TagSpec('uv', 14, 16, '')],
+  //         new TagSpec('p', 19, 20, ''),
+  //       ],
+  //       groupSpec: new GroupSpec(0, 20, [new GroupSpec(3, 7, []), new GroupSpec(9, 17, [])]),
+  //     },
+  //     {
+  //       name: 'tag-after-group',
+  //       string: 'x/y,(r,v)',
+  //       explanation: 'A tag after a group.',
+  //       tagSpecs: [new TagSpec('x/y', 0, 3, ''), [new TagSpec('r', 5, 6, ''), new TagSpec('v', 7, 8, '')]],
+  //       groupSpec: new GroupSpec(0, 9, [new GroupSpec(4, 9, [])]),
+  //     },
+  //   ],
+  // }
 ]
